Switched from libffi::low to libffi::middle
libffi::middle takes care of some memory management and wraps some
function calls in a more safe-ish manner. Libffi-rs also has a "high"
interfeace, but I have currently not been able to find a way that can be
used for this project.

Also replaced the ffitype_to_native function to to_libffi_type, which
is implemented on the FfiType enum.