Added sqlx-data.json for "offline" compilation
The sqlx-data.json file and offline compilation made cargo behave
better in certain instances, such as when doing tests. However, in order
to do this, sqlx-data.json needs to be rebuit when the database schema
changes (or new queries are written/old ones modified?). Added a section
in the readme about this.

Also fixed formatting for a lot of files.