Started work on HTTP client using Win32 API
rustrat-client currently contains a poc that is able to send a GET
request and return a Vec<u8> of the URL. The next step will be to
attempt to rustify the code and make it usable for C2 communication.

Added a new verb to rustrat-client-exe.exe, `http`. This will attempt to
send a GET request to www.wodinaz.com and print the returned source.