First commit with badtui
badtui is the TUI that I will use for the server until version 1.0,
which should include a web-based interface. Currently there are some
issues, most likely with sqlite when starting the server. I might not
fix it before version 1.0 if I am able to live with it.

Currently, the TUI is used as the output for logging as well, this is
not ideal either, but I will see whether it is something I am able to
live with for now.